Redefined the bar colors for mass limit and CPU limit as green, red, orange
These bars are purely aesthetic and don't need to be colorblind-accessible,
so I figured that keeping green, orange, and red as the colors is
reasonable. It looks nice and that's all that matters for these bars.

Also removed the cDRestricted color, which is no longer used or needed.
Just for you, @bobbens. 😛