Improved colorblind shader, added an in-game "colorblind mode" option.
The in-game "colorblind mode" option is more limited than editing the
shader directly as before, but the menu option enables using
colorblind mode for users. This is useful in particular for those
with blue cone monochromacy if they are playing in high-light
conditions, since some things are hard to see with the short cone
cells that are active for them in bright conditions. It is also useful
for someone using a red-shifter (like Redshift or f.lux) with a very
low color temperature, as for these users blue pixels can turn black
otherwise. Technically speaking it's also "night vision", which some
users might like using just for the heck of it.

The script now uses #defines for exact colorblind mode rather than
just magic numbers, but the commented out #define used to enable
colorblind modes other than rod monochromacy still contains a magic
number since it's easier to edit on-the-fly.