Made double-clicking on jump points engage autonav.
Not sure what the purpose of the previous behavior was honestly.
It just caused you to jump *if* you were already in position and
able to jump, without making any moves to bring you into said
position (the same as the J key I believe). It never struck me as
something you would actually need to do.