Started on the new stealth system (this is a bit of a mess)
So, here's the rundown: the system currently works fully for visibility,
but no outfits have been given the new stats to work with this yet,
and systems haven't been given the replacement for "interference"
yet, so everyone can see everyone else at a range of 7500 right now.

Weapons tracking is still stuck to the old system, though a new function
has been added for the new system.

Most of the effects of the "interference" attribute of systems have
been removed, with the exception of those that affect weapons. This
includes the interference effect on the GUI that faded in and out,
which I've thought for a long time is both an accessibility problem
and doesn't add anything, anyway. So that effect is gone now.

Next steps:

* I want to automatically set rdr_range_mod values of systems based
  current "interference" values, where the rdr_range_mod will become
  -interference / 10. Having looked at the numbers, I think this will
  be an okay first approximation for the effect we want. We can then
  adjust the numbers later.
* Weapons of course need to be brought up to the new system. This will
  require creating some now attributes for them, I think.

I think most likely what I'll do first is focus on bringing the
detection aspect up to standard before touching the weapons. That
way I will have a properly functioning test environment sooner.