Started adding a few pieces for a new GUI called "Neo".
The concept for Neo is for it to be a simple GUI taking some of the
visual design of Brushed, the functionality and basic layout of Slim,
and the simplicity of Legacy. Starting this from the ground up because
none of the existing GUIs are all that close to the design I'm going
for here.