Use %LG for display of population.
Population is a figure that doesn't matter for the game, and so it's
purely cosmetic, meaning that the way it's displayed needs to look
nice and doesn't need to be accessible. While many may not understand
exponential notation, it avoids giving the impression that even
numbers like 4 billion are exactly how many people are there. Also,
exponential notation looks cool.