Only display stats that are relevant.
This has always been somewhat of a problem, but exacerbated by the
recent change of several values to ship stats: often, an individual
ship stat just isn't relevant at all when looking at a fully
equipped ship. This change adds a feature to the shipstats system
to mark which stats should be omitted in that circumstance and
adds an argument to ss_statsDesc to make use of the feature.
Currently used only by equipment_shipStats, which seems to be
sufficient.