Changed the board window to simplify it and make its text bigger.
I decided to change it so that the buttons also serve double duty
as headers for the info. This required reorganizing things so that
each stat had its own label. However, I think it's a much better
approach.

This also adds back in an old feature to see the type of cargo,
which was removed because it caused bugs due to wrapping. Cargo
has been moved to the bottom, and it is given space to wrap into
the empty line below.