Boarding window: added more spacing: allow ammo text to wrap.
This expands spacing in the boarding window such that it's now possible
for text in the ammo portion to wrap as the cargo portion can, albeit
to a lesser extent. This will almost never happen, but just in case,
that works now. Testing this, it's able to hold 6 ammo names comfortably,
which should give a decent buffer.