Un-nerfed the practice drone.
It was originally nerfed before the rebalance. Now, the player has to
do a very very bad job to be in a situation where they're unable to
kill the drone with standard cores, and the low-end cores are so weak
that the player almost certainly won't even see their heat meter going
up and the active cooldown segment might even trigger without them going
into active cooldown, which is a big no-no.

Also removed the text talking about active cooldown before the fight,
since it's to be taught in the next section and it's best for the player
to allow heat to build up so they can see its effects.