Removed that clunky debug message of planet_getSystem.
Since much of the code was implicitly expecting a NULL value and
responding to it in various ways, and in fact fundamental aspects
of the game depend on a planet which isn't in a system being valid,
I've removed that silly debug message and instead gone through and
made sure that *all* usages of this function properly account for
a NULL return. Also explicitly documented the NULL return case.