Added a feature to allow players to "dock" pilots.
This feature simply causes a pilot to not spawn in space, while
still retaining royalty payments to them. This allows leaving
pilots sort of "on retainer" if you want to do a mission which doesn't
require them, or even allows you to rotate pilots if you want to
do that.