Buffed Particle Beam's turn rate.
This makes it very easy for the Particle Beam to keep up with the
movements of fast ships, making up for its short range and low damage.
It can now be used in large numbers by a dedicated Vendetta pilot to
easily destroy ultralight ships that get too close (though it is still
a highly unbalanced and generally not worthwhile weapon).