Spawn a bunch of Empire ships for comingout4 to make it easier.
This is the most unobtrusive way I could think of to make this
mission less incredibly difficult. Since Chelsea's ship couldn't
be made faster due to it being an escort mission and was already
as strong as a Llama possibly could be, the only way to make it
easier is to add friendly ships or remove hostile ships. The latter
isn't easy to do, so I instead opted to absolutely fill the systems
traveled in with Empire ships, if the player is at least neutral
with them. This helps to keep the pirates in the systems busy.