Removed traffic_00.
I was thinking of maybe fixing this so that it didn't cause conflicts
with the core message of the game, but given the complexities of what
actually causes human trafficking, I don't think it can be addressed
properly.

The problem with the mission as it stands is that it suggests that
human trafficking is caused by a select group of "criminal scum",
which is heavily over-simplifying a serious issue. The fact of the
matter is that human trafficking is a very real thing that usually
involves powerful rich people, but is carried out by standard grunts
not unlike pirates. Human trafficking is also far more complicated
than just kidnapping children and I don't even think most cases of
human trafficking are kidnappings; far more common is situations
where human trafficking is a way to exploit people trying to get
out of a dangerous situation in a poor or totalitarian country,
which itself is largely fueled by a complex system of imperialism.

There is no way I could ever do this topic justice in a side-quest
for a game, and the existing mission doesn't even come close,
suggesting somehow that if you just get rid of "the human traffickers"
then human trafficking will go away. Not to mention, the setup of a
random pirate kidnapping children to traffick is wholly unrealistic.
If anything, it would make much more sense to cover the topic in the
Runaway campaign, which involves a runaway who would absolutely be
put in a position where they would be vulnerable to exploitation, but
that campaign is already finished and didn't go in that direction.

So yeah, I'm just abandoning the idea. It's a good idea for a topic
in principle, but it simply cannot take this simplified form.