Display all services, made map bigger to accommodate this.
Previously it was only showing certain services to show on the
star map which were a bit arbitrary. Now it shows every service
except for "Inhabited", since that's not ever useful to a player
(it basically only means for a player that ships won't spawn from
it and it's not allowed to have 0 population).

I also needed to adjust the ordering of services to include the Land
service while still excluding the Inhabited service. On the way, I
notice that the <uninhabited/> tag had been removed from Domestica
and that Foundation Station was missing the Black Market service,
so I fixed those.