Removed a broken feature from pilot.addOutfit.
I didn't even realize this was there, but there was an attempt to
allow outfits to be implicitly overwritten with pilot.addOutfit if
the outfit was a default outfit. This didn't actually work and in
fact was messing up the equip script when attempting to equip outfits
that happened to be default outfits. Absolutely a mess, and not even
a good idea.

This commit also adjusts whitespace in the equip script and slightly
adjusts the Sindbad "ship" (giving it more energy and slight armor
regeneration).