Touched up Empire shipping, fixed boarding tutorial/applied to es02.
This makes the entirety of the long distance cargo campaign required
to start the Soldner shipping campaign, and also fixes formatting to
be double line breaks and slightly tweaks dialog. The VIP mission has
also been given the boarding tutorial, since you have to board a ship
in that mission. While I was doing this, I noticed that the other
boarding tutorials were buggy and fixed them.