Added a display to show what services are actually available.
This improves the services list by giving the services colors and
symbols corresponding with the planets they are on. The most lenient
service availability is used for this coloration, so e.g. if a
restricted asset and unrestricted neutral asset both have an outfitter,
the outfitter services is displayed with unrestricted neutral color
and symbol. This display allows you to see when a service isn't
actually available because it's only on an asset you can't land on,
which is useful in systems with multiple factions or systems with
some restricted assets and some unrestricted assets.