Dogfight aiming helper improved and enabled by default.
The dogfight aiming helper now chooses its margin-of-error guides
based on what weapons you have equipped that are in range,
auto-aligns if all your weapons that are in range are turrets, and
only shows the aim cross if at least one of your weapons is in range.
The effect of this is that the aiming helper is far more flexible and
also shows whether or not you're able to hit the enemy at all.