Fixed slightly broken equipment swapping code.
It caused the wrong outfit to be added after removing the outfit if
the size of the image array changed. Come to think of it, I'm not
entirely sure why because `o` is a pointer. Perhaps some sort of
compiler optimization and the way the memory is stored, or maybe
I'm just too tired to understand right now.

Either way, this adjustment fixes it, even if I don't fully
understand why.