Added confiscation of cargo and outfits to waste dump fine
We realized that it's possible and quite easy to work around the
fine for aborting while landed just by buying a bunch of outfits
before aborting, so we decided to make this much more difficult and
thus discourage it by causing all cargo to be confiscated to apply
to the fine, and then causing outfits to be confiscated until the
fine can be paid for, or there are no outfits left.

For simplicity and because working around this using ships would be
more trouble than it's worth anyway, ships are not considered, so
it's still possible to sneak around the restrictions by using ships
as sort of storage for outfits, and buying ships to sell after the
illegal dump. This is a lot more setup, though, so we think this is
fine. As long as the exploit is as difficult as just going to one of
the proper destinations or dumping in Arandon, it's fine that it's
possible.