Made several changes to improve the space family mission.
This makes it so that it's possible to retake the mission if you
abort it. It also fixes up the abort text, and removes the Gawain
from the list of possible ships as that is the one which is most
likely to be accidentally destroyed (and doesn't make sense as a
trader vessel).