Removed Particle Beam.
My experimentation shows that the Particle Beam is absolutely,
unequivocally useless. Damage output is fine, but for the smallest
ships, it's incredibly important to have better range than your
opponent, and so trading so much range for extra damage will never
make sense. In order for it to be even remotely useful, it would
have to have longer range than the Orion Beam, which would be very
odd.

Beams work great when combined with missiles for long-range combat.
However, the ships that would use the Particle Beam don't have
missiles to combine it with. The only rockets available at all are
forward-mounted, and depending on forward-mounted rockets kind of
defeats the purpose of using a turret beam.