Set player weapons to only fire if in range by default.
Trying to fight without targeting enemies is never going to be
optimal, and weapons simply can't work as well if they're allowed
to fire when not in range. The only advantage of not setting inrange
to 1 by default is that new players can see without targeting
something what buttons to use to shoot. Since the tutorial explains
how to target enemies, that's not worth the long-term annoyance of
having to manually set weapons to not fire out of range every time.