Removed the boring shipping diplomat missions, re-adjusted es00 reqs.
This makes the es_longdistancecargo mission available immediately on
joining the Empire, rather than making it a reward for delivering 6
diplomats, which was an extremely tedious campaign. Since those
missions are gone now, es00 is now back to roughly its original
requirements, with one adjustment: it requires Empire rep of 10 or
more instead of 0 or more.