Removed some uses of pilot.addFleet, expanded addShips API to help.
I made it so that you can pass a table of counts rather than a
single count to addShips. This helps with using it to spawn fleets
containing varying amounts of different ships without having to
write tables with lots of duplicates, while also keeping the
effect of the previous single argument (only difference is the order
the ships show up in the table returned).