Added further safeguards for 0 fuel consumption.
This makes everything fully able to handle 0 fuel consumption,
including allowing player.jumps() to return nil for infinite jumps
(and handling that API change in existing usage of the function).
This is currently completely unnecessary, but it safeguards just
in case we want to add something with infinite jumps at some point
for some reason (or e.g. set 0 fuel consumption for testing).