Removed Crimelord mission.
This wasn't altogether a terrible idea for a mission, but didn't
really make much sense as a Dvaered space mission and sends a
political message I'm not fond of. It also is clear from internal
comments that this was largely a test mission. Given that, I think
I'll remove this mission and think about adding a new similar
mission in the future either as a part of a campaign (FLF?) or just
done in a more generic way and with a different (not pro-cop) message.