Added a check to ensure no infinite loop is possible.
The infinite loop was not at all likely, but it was technically
possible if the player had so many ships with the naming scheme
"temp-X" that it took more than 9 characters ti write the digits
out, which would cause the number to increase infinitely and the
string to never become unique (leading to an infinite loop). This
also silences a warning about the possibility.