Fixed refueling so that it always refuels in amounts that makes sense.
This turned out to be pretty easy on the surface, but I had to make
one other change in tandem with that: the way fuel was loaded onto
ships. It was awful and led to uneven fuel amounts. As such, I've
changed that so that it's actually specified by the equip script and
always fuels in multiples of fuel consumption. For ships that don't
use the equip script, they will end up with max fuel (given default
cores), e.g. 300 for a Llama and 1100 for a Schroedinger.

Now, as for the actual goal of the commit, this makes it so that
the code now attempts to give the player one jump's worth of fuel if
possible. If it can't do that without leaving at least 1 jump of
fuel for the donor, it instead falls back to giving the amount of
fuel − the fuel consumption (leaving the donor with 1 jump). This
also entailed modifying how the player is made aware of the amount
of fuel gained; instead of the occasional hardcoded "100 hL", the
C side now tells you exactly how much fuel you're getting. (I chose
to do it on the C side because it's simpler that way and ensures
consistency, given that amount given is calculated on the C side.)

This change makes refueling from other pilots much less of a hassle
than before if you have a large ship. While you likely will still
need to get fuel from several ships, you will at least get all the
fuel you can from each ship with one refuel (rather than having to
ask the ship for fuel over and over again).

This commit also includes a minor edit to the readme that I forgot
to commit before, adding physfs to the list of recommended Ubuntu
packages.