Started importing Naev's ai.core module with Naikari-specific changes.
I don't know for sure if I'll finish this. Right now, it's not in
a working state as it still references safe lanes stuff (which Naikari
doesn't have), and I believe there are some Naikari AI features that
don't exist in Naev's code, so I'll need to port those over. In any
case, these files are out of the way and currently unused. If I finish
this, I can switch all the actual AI files to use it and make Naikari's
code base in sync with Naev's.

Main issue is deciding what to do in the places where the safelanes
system is being used to guide loitering in Naev. It's only a couple
places so it shouldn't be too terribly hard, but I've got a bit of
a headache right now and am a bit exhausted, so I'll hopefully finish
up later.