Improved Seek and Destroy mission in several ways.
This includes major code cleanup, removal of window titles, and
removal of the use of pilot.addFleet.

This also makes a couple more substantive changes to the mission.
Firstly, the ambush mechanic was removed; it was weird, possibly
frustrating, and didn't really work all that well considering the
ambush ships would often just tell you where to go next. Secondly,
I gave the FLF a particular characteristic of being noble and
fearless. An FLF ship will never, ever rat out a fellow FLF pilot.

Because I don't like it aesthetically, I also removed the comm
messages which were shown when you bribe or attack a ship for
information. These were redundant and had the potential to not
really fit with the tone of the previous message which already told
you that your threat didn't work or that your bribe did work.