Added macros to make weapon set magic numbers named.
This is the first step of what I hope will be a conversion to
using reserved AI-exclusive weapon sets for the special meanings
meant for AI. The hope here is that this will lead to more
flexibility in designing things around how a *player's* weapons
should be configured by default, while also making it possible for
the configuration for AI to have more flexibility.

A PLAYER_WEAPON_SETS macro has been added which is currently the
same as PILOT_WEAPON_SETS. It will remain at this value (10) even
after PILOT_WEAPON_SETS gets increased, and is used to determine
what weapon sets to show to the player.

And for the weapon sets themselves, macros have been added indicating
their name.

The leftover weapon sets have been left alone for now since what I
plan to do for those is to just have a for loop setting all of them
to WEAPSET_TYPE_ACTIVE by default.