Removed the new-style AI stuff.
It's just too different I think. I would love to be able to
incorporate some of its improvements but it's just far too closely
connected to the whole lanes thing. I'll retain the old API and try
to backport improvements to it when I can, as I did in the last
commit.