Added support for a weapon set indicator string in some Lua functions.
Used it in a couple places (haven't replaced all applicable numbers
with strings yet) to enable testing. Took some trial and error, but
it seems to work now. Next step will be making use of these strings
everywhere it makes sense to (basically just the AI code) and see if
I missed any Lua functions which should have this functionality,
which is not unlikely. Once all that's done, we can start changing
the weapset numbers gracefully.