Automatically recall fighters rather than complaining about it.
This causes the actions that previously were simply forbidden with
warnings telling you to recall your fighters to now automatically
undeploy fighters for you. Swapping a ship (either with a purchase
or just to choose another of yours) recalls all deployed fighters,
and unequipping a fighter bay causes all of its fighters,
specifically, to be undeployed.