Converted the little info on the land screen to a single text widget.
This is something I'm planning on applying all across the board. The
whole header/data differential has always been a major pain when it
comes to things like translations because of creating space
constraints. What prompted me to do this is noticing that the space
constraints didn't leave enough room for a dyslexia font, and so
since I'd like to add support for one, some changes are needed.

These changes will be applied to everything else later. Taking a
break for now.