Removed rehab missions.
Rehab missions are not something in-line with what we want because
it ends up requiring unreasonably tight controls on how reputation
can be gained to some degree to avoid making it fairly easy for a
player to just be friends with every single faction in the universe.
In particular, eliminating rehab missions is necessary to eliminate
the pirate fame drop event without serious consequences.