Changed the common rep-boosting missions to use modPlayer.
This ensures that these missions cause a rep loss for their enemies,
most notably, pirates (but this also affects the FLF as well in the
case of Dvaered missions).

We may want to consider some mechanism to still make it easy to not
become enemies with the FLF. Then again, maybe it makes sense that
the FLF would be suspicious of mercenaries that do jobs for the FLF,
even if those jobs are always hunting pirates.