Added more of the needed text to flf_dvk02. Almost done!
Alright, after quite some time trying to figure this out, I've got
a good solution. The player goes to Fury Station and uses a carrot
and stick approach, threatening destruction but offering bountiful
rewards if they cooperate. Of course, the player does this with no
knowledge of the nature of how they will be involved, only offering
involvement and riches beyond their wildest dreams. This then solves
the pirate problem and makes the next task living up to the "riches"
promise. It will eventually lead to the FLF's downfall.