Removed the Dvaered "Diversion" mission.
I don't think this old mission is really worthwhile, having looked
at it, and it hasn't been touched in 9 years. It's a mission that's
so rare to run into that I literally have never seen it activate ever
in the years I've been aware of Naev, because it only works when
you land on a very specific planet that most players would have no
particular reason to land on.