Changed the player radar color to a light version of the friendly color.
The purple color is far too similar to the hostile color, which is
a bit odd. Making it a lighter version of the friendly color allows
it to be more visually similar to friendly ships, which makes more
sense. The only downside is that rod monochromacy leads to it looking
more like the neutral color, but that's not a terrible downside since
the player graphic is distinct anyway.