Lowered torpedo speeds, increased radar ranges.
This makes all torpedoes much more in-line with the potential to
reach targets they're supposed to have. Before, it was far too easy
for e.g. a Caesar IV to chase down a Pacifier. Caesar IVs are now
just barely fast enough to be able to overtake the fastest Cruiser
(a Kestrel with a Krain Remige), but not quite fast enough to
overtake the slowest Destroyer (one with a Unicorp engine).

This also significantly increases maximum radar range for all
torpedoes because it allows for more of a buffer to exist, making it
more possible for these missiles to be used in nebulae or at the edge
of a ship's radar range.