Added a function for equipping player's stored ships.
player.shipOutfitAdd adds a major missing functionality for interacting
with the player's ships, something that will be needed by the upgraded
ship stealing event. Originally I tried implementing a way to grab
the other player pilots, but this proved to be impossible without
major architectural changes, so this complement to player.shipOutfits
has been added instead.