Added a new ship stealing mission and fixed some Assassination Job problems.
The new ship stealing mission isn't tested yet but should be much
better and more interesting than the previous one. Instead of simply
landing on a planet, you actually locate and battle the ship, and the
pirate informer pilots the new ship out of the system. You are given
the new ship with all the outfits it had while you fought it.

While I was doing this, I noticed some problematic code in the
Assassination Job mission, most significantly a part where overly
aggressive find-and-replace had made the mission guaranteed to fail
when accepted, so that's fixed. Also, the old ship stealing mission
has of course been removed.