Made rush cargo missions more lenient with jumps per stop.
I don't know why it was this way, but the rush cargo mission was
coded with the expectation of needing as many as 6 jumps per stop
of fuel for the higher tier missions. There is no reason whatsoever
to expect a courier ship to have 6 jumps worth of fuel, so it's no
wonder they were so unforgiving to early ships. This changes the
range of values so that it's from 1 to 4 instead of from 3 to 6.