Gave the Llama a very slight speed bonus.
The speed bonus is the same one given to the Hyena, allowing the
Llama to perform less badly against Hyenas (which should help
early game when the Llama is all you have). It still lacks the
maneuverability and is not as fast as the Gawain or Quicksilver,
but this should at least give the Llama more of a fighting chance
against pirates.