Increased Gawain's speed further and added some nerfs to compensate.
The main thing I'm concerned with balance-wise when it comes to
increasing the Gawain's speed is a player's ability to use it like
an ultralight fighter by taking advantage of rockets. To that end,
I've applied nerfs to the ship's capability to use rockets. They're
not entirely useless, but the reduced ammo capacity and damage makes
it nearly impossible to destroy a ship alone using missiles.