Removed unused ship stats and cleaned up display of a couple.
Two stats have been removes: tur_tracking (which did nothing) and
launch_lockon. Now, launch_lockon did do something, but it was a
behavior that I don't think anyone would ever expect: it increased
the amount of leniency time allowed to keep a lock-on after turning
away from a target. That's something that would be useful and even
appreciated, but would be almost impossible to properly teach the
player about it without a specialized tutorial or something, which
is silly for something with such a small effect.