Simplified the dialog width chooser.
The previous method was… weird, and didn't make a whole lot of sense.
I think it was deciding how much width is "needed" or whatever based
on how much total content there was or something. But this is rather
silly, it leads to unnecessarily slender dialog boxes and I don't see
the point in that. This new, simpler method instead just makes the
dialogs narrower if the whole content can fit into a narrower space
without line breaks. Makes a whole lot more sense to me than whatever
it was doing, and means that large dialogs with paragraphs will always
have the max 800px width.