Shrunk down Brushed GUI's bottom bar to make it less obtrusive.
The reason I'm doing it this way instead of getting rid of it entirely
is because it prevents the graphics above it from looking like they
randomly cut off. Still not 100% perfect of course, but better than
before and since Brushed will be replaced by Neo, a bit of imperfection
is fine.