Scrubbed all uses of the term "thug".
The term "thug" originates from "thuggee", which may be either
made-up, exaggerated by propaganda, or caused by the British occupiers
of India, according to various accounts. While I'm not sure of it,
switching to "gangster", "mercenary", or "goon" (as appropriate)
avoids the issue, and these terms work just as well. Besides, even
if the British account of thuggee is accurate, the term "thug" is
often used today in a racist manner, something we don't want to
associate with.