Slightly simplified and touched up the Za'lek cargo run mission.
This in particular cuts the part where the comm window is opened,
since the gag of "look, I got my ship back!" doesn't apply. Also
uses the fmt library instead of numstring and uses fmt.f for one of
the strings (not retrofitting the rest right now since it's not that
high of a priority).