Added an invincibility event for brief invincibility after takeoff.
This is a simple event: it hooks takeoff and when that happens,
makes the player invincible for 3 seconds, just in case the player
spawns right on top of an enemy firing at them. The event also offers
an override option for missions and events in case it can cause
problems.