Improved the way the player being in the process of hyperspacing is handled.
This change adds two particular improvements. Firstly, a check is
added during hyperspacing to see if autonav should reset speed,
avoiding speeding up the game while enemies are firing at you.
Secondly, and more importantly, this ensures that it's possible to
cancel the jump sequence just by tapping any movement control.
This is not done for the autoface key under the logic that autoface
should face you toward hyperspace, and is not done with the weapon
keys so that this doesn't leave you sitting there wondering why
you're not hyperspacing simply because you were shooting at someone
when the sequence started. It is also not done with mouse control
since that would be more frustrating than helpful.