Added a "secondary" standing hook argument, improved Fake ID and pir cargo.
Ok, this is a bit of a mess. I was originally going to use the new
"secondary" argument to help out the Fake ID mission, but then
realized that it wouldn't help. Still kept that feature in, and
for what I wanted to do with the fake ID mission, I made the pir_cargo
mission use modPlayerSingle if the fake ID mission is active. While
I was at it, I touched up some internals and also added a possibility
to get your identity discovered if you gain more than 20 reputation
with one of the factions who you were wanted by.