Made it possible for auto-landing to include bribing in its sequence.
This took me hours to figure out how to do, but now that it's there,
it works well. It now waits for you to get in-range, and once you do,
automatically opens a hail window. If you bribe, autonav continues.
If you don't bribe, autonav is aborted. A nice side-effect is that
the same infrastructure could be used to inform the player of
LANDACK before actually reaching the planet if initial contact was
impossible. Overall, a considerable improvement.