Added a bit of a lampshade on the horribly badly scaled distances.
Truth is, Naikari distances can never possibly be even remotely
reasonable. Using megameters for distance would be closer, but still
not right, and it would cause ships to be traveling faster
than the speed of light. So instead, adding an NPC lore text that
explains this away as "miniature dark matter wormholes" found by
"quantum triangulation", with a mention of exactly how off these
distances are. (With the figure given, the true size of Sol in-game
is about 23.4 au, compared to the 30 au distance from the Sun to
the inner edge of the Kuiper belt.)