Accurately calculate if ship will hit jump for instant jump.
For the quicker method to instant-jump, a lazy formula was used
previously where it just picked a flat angle. It now actually
computes the forgiveness angle it should be using based on the
distance to the jump and the jump's radius, making it use the
better method in every case where it is more efficient to do so
without causing it to be attempted in cases where it isn't ideal.