Made pir_cargo and es_cargo more time-lenient with jumps and static.
These missions don't show what "tier" they are, so having some of
them doable with a starter Llama and some not could be very confusing
and annoying to deal with. These missions are supposed to be quite
generous with their time limits, so I've made it so. It now gives a
standard 15000 seconds (STU) per jump and a standard 3 jumps per stop.