Restored accidentally removed function, changed how fmt.credits works.
fmt.credits now works as credits2str does in the C API, which avoids
the issue of needing to update all references to credits to use
ngettext since that method shows decimal amounts (which thus shouldn't
be passed to ntgettext anyway).

While I was doing that, I noticed that the loads function (used by
fmt.f) was accidentally removed, so restored that as well.