Fixed up the odd, essentially broken implementation of system.jumpDist.
I don't know why, but the jumpDist function would return 0 if there
was no route available with the given parameters. Not only that, but
it rejected nil as a valid argument for parameter 2 even though it
said it was valid in the documentation. Quite a mess. I've reworked
it so that it returns nil, not 0, if there is no route, and adjusted
a few missions that will throw an error if given nil so that now
they won't. What a headache (literally)…