Fixed a couple issues with cargo missions.
Firstly, due to a typo cargo missions weren't showing jump distance.
Oops!

Secondly, a special option has been added to getsysatdistance to
take into account the existence of hypergates. Pirate missions were
getting targets all over the galaxy due to a combination of their
long range and the fact that they do go through hypergates. This
avoids that. Of course, when hypergates enter the story, it should
be very possible to pass a "true" argument to that hew "hypergates"
parameter if we decide to do that.