Use modPlayerSingle for the rep loss in flf_catastrophe.
Using faction.modPlayer() with a negative amount poses a risk of
creating avenues to raise reputation with factions that the player
should not be gaining reputation with. In the case of doing this
with the FLF, a standing gain would apply to the Empire, the Dvaereds,
the Collective, and the Pirates.